Heidy rama

+++ front_procrastinator/src/app/user/index/index.component.ts
@@ -1,27 +1,35 @@
 import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { UserService } from '../../servicios/user.service';
+import { RolService } from '../../servicios/rol.service';
 import { Router } from '@angular/router';
 import { User } from '../../modelos/user.model';
+import { Rol } from '../../modelos/rol.model';
+import { ActivatedRoute } from '@angular/router';
+import Swal from 'sweetalert2'
 
 @Component({
   selector: 'app-index',
   standalone: true,
   imports: [CommonModule],
-  providers: [UserService],
+  providers: [UserService, RolService],
   templateUrl: './index.component.html',
   styleUrl: './index.component.scss'
 })
 export class IndexComponent {
+listaRol: Rol[] = [];
 listaUsuarios: User[] = [];
 clave: string | null = null;
 usuario: User | null = null;
+id: string | null;
 
-  constructor (private userService: UserService, private _router: Router){}
+  constructor (private userService: UserService, private rolService: RolService ,private _router: Router, private aRoute: ActivatedRoute){
+    this.id = this.aRoute.snapshot.paramMap.get('id')}
 
   ngOnInit(): void{
     this.validartoken();
     this.cargarUsuarios();
+    this.cargarRol();
     }
 
     ngOnchanges(): void{
@@ -45,16 +53,55 @@ usuario: User | null = null;
         console.log(err);
       });
   }
+
+  cargarRol():void{
+    this.rolService.getRol(this.clave).subscribe(
+      data1 =>{
+        this.listaRol = data1;
+      },
+      err => {
+        console.log(err);
+      });
+  }
+
+  mensajeSiono(text: string, deleteText: string, id: any, confirmButtonText?: string, timer?: number) {
+    Swal.fire({
+      title: "¿Estás seguro de eliminar el usuario?",
+      text: "¡No podrás revertir este cambio!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí"
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          title: "Borrado",
+          text: "El usuario ha sido borrado",
+          icon: "success"
+        });
+          this.eliminarUser(id);
+      }
+    });
+  }
   
   eliminarUser(id:any): void{
     console.log(id);
-    this.userService.deleteUsuario(id, this.clave).subscribe(
-      data => {
-        this.cargarUsuarios();
-    },
-    err => {
-      console.log(err);
-      });
+    if (id >= 2){
+      this.userService.deleteUsuario(id, this.clave).subscribe(
+        data => {
+          this.cargarUsuarios();
+      },
+      err => {
+        console.log(err);
+        });
+    }else{
+      console.log("no puede eliminar el usuario administrador")
+    }
+    
     }
 
+    verApps(id:any): void{
+      this._router.navigateByUrl("/app/index/"+id);
+  }
 }
